+++ mink/core/user.py
@@ -13,6 +13,8 @@ class User:
         id: str,  # noqa: A002
         name: str,
         email: str,
+        idp: str | None = None,
+        sub: str | None = None,
         ui_language: str | None = None,
     ) -> None:
         """Init user by setting class variables.
@@ -21,11 +23,15 @@ class User:
             id: The user ID.
             name: The user's name.
             email: The user's email.
+            idp: The user's identity provider.
+            sub: The user's subject identifier from the identity provider.
             ui_language: The user's UI language.
         """
         self.id = id
         self.name = name
         self.email = email
+        self.idp = idp or ""
+        self.sub = sub or ""
         self.ui_language = ui_language or "swe"
 
     def __str__(self) -> str:
@@ -34,7 +40,14 @@ class User:
 
     def serialize(self, depth: int | None = None) -> dict:
         """Return a serialized dict representation of the user instance."""
-        raw = {"id": self.id, "name": self.name, "email": self.email, "ui_language": self.ui_language}
+        raw = {
+            "id": self.id,
+            "name": self.name,
+            "email": self.email,
+            "idp": self.idp,
+            "sub": self.sub,
+            "ui_language": self.ui_language,
+        }
         return utils.serialize_obj(raw, depth=depth)
 
     def set_parent(self, parent: Any) -> None: